finish shapefile convert to wkb
